backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
build==1.3.0
CacheControl==0.14.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
cleo==2.1.0
click==8.3.0
comm==0.2.2
crashtest==0.4.1
debugpy==1.8.5
decorator==5.1.1
distlib==0.4.0
Django==5.1.1
django-crispy-forms==1.8.0
django-formtools==1.0
django-tables2==2.1.1
dulwich==0.24.2
et_xmlfile==2.0.0
executing==2.1.0
fastapi==0.104.1
fastjsonschema==2.21.2
filelock==3.19.1
findpython==0.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
installer==0.7.0
ipykernel==6.29.5
ipython==8.27.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jedi==0.19.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==25.6.0
matplotlib-inline==0.1.7
more-itertools==10.8.0
msgpack==1.1.2
nest-asyncio==1.6.0
numpy==1.26.4
openpyxl==3.1.2
packaging==25.0
pandas==2.1.3
parso==0.8.4
pbs-installer==2025.10.7
pexpect==4.9.0
pkginfo==1.12.1.2
platformdirs==4.3.6
poetry==2.2.1
poetry-core==2.2.1
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
pydantic==2.12.1
pydantic_core==2.41.3
Pygments==2.18.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==26.2.0
RapidFuzz==3.14.1
requests==2.32.5
requests-toolbelt==1.0.0
setuptools==75.8.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.27.0
tomlkit==0.13.3
tornado==6.4.1
traitlets==5.14.3
trove-classifiers==2025.9.11.17
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
virtualenv==20.34.0
wcwidth==0.2.13
xattr==1.2.0
zstandard==0.25.0
openai==1.3.7
python-dotenv==1.0.0
